can create basic generic client for a few providers
