+++ docs/gitea/gitea.yaml
