--- generic/README.md
@@ -1,13 +1,0 @@
-# Just a generic git provider based on go-git-providers
-
-# Gitea 
-
-
-- I could install  via tmp/gitea.yaml
-
-![](docs/Screenshot 2022-11-01 at 17.41.19.png)
-
-- I could create a client via test
-- I could not create PR due to create branch not avaialble
-  https://github.com/jenkins-x/go-scm/blob/main/scm/driver/gitea/git.go#L40
-
